Update README.md
This project is based on the python it helps in the enrolling of the job for a fresher. Skilled Python developer with expertise in creating user-friendly GUI interfaces. Demonstrated proficiency in designing and implementing the "Enroll for a Job" project, incorporating database connectivity. Proven ability to enhance user experiences by seamlessly integrating job application functionalities, interview question access, and employee details for improved job-seeking processes.